refactor(mysql): extract CLI subcommands into modular cli/ subpackage
Split the monolithic __main__.py (~1611 lines) into focused subcommand
modules under cli/:

- __init__.py: subcommand registration and dynamic handler dispatch
- output.py: output provider factory and display helpers
- connection.py: connection argument parsing, backend creation, config
- info.py: MySQL environment info with protocol support display
- query.py: SQL execution with sync/async support and --log-level
- introspect.py: database introspection with sync/async support
- status.py: server status overview with sync/async support
- named_query.py: named-query adapter for shared CLI helper
- named_procedure.py: named-procedure adapter for shared CLI helper
- named_connection.py: named-connection adapter (no -o, no connection args)

Each subcommand now independently decides its own arguments:
- info: connection args, -o {table,json}, --verbose, --rich-ascii, --version
- query: connection args, -o {table,json,csv,tsv}, --log-level, --rich-ascii
- introspect: connection args, -o {table,json,csv,tsv}, --rich-ascii, --schema
- status: connection args, -o {table,json,csv,tsv}, --verbose, --rich-ascii
- named-query/named-procedure: connection parent parser, -o, --rich-ascii
- named-connection: empty parent parser (no -o, no connection args)
